new libary, clipboard-copy now supported
The generated code, can now be copied out of the browser thanks to the libary clipboard. This lib was used, because as with many things in HTML, you would have needed to implement that feature for every browser (sometimes even version) differently. So the library took away a lot of work...